Split up admin into seperate commands
